create UniversalProjectsRoute and move invite code outside of it
create UniversalProjectsRoute and use it for all routes
move transfer modal and banner code outside MyProjectsRoute
