+++ jsapp/js/components/permissions/transferProjects/inviteBanner.module.scss
@@ -1,0 +1,19 @@
+@use 'scss/colors';
+
+.banner {
+  display: flex;
+  justify-content: space-between;
+  margin: 24px 24px 0;
+  padding: 12px;
+  background-color: colors.$kobo-light-blue;
+  border-radius: 5px;
+  align-items: center;
+}
+
+.bannerIcon {
+  padding-right: 18px;
+}
+
+.bannerButton {
+  margin-left: auto;
+}
