+++ jsapp/js/projects/projectViews.module.scss
@@ -1,5 +1,7 @@
+@use 'scss/z-indexes';
 @use 'scss/mixins';
 @use 'scss/colors';
+@use 'sass:color';
 
 .root {
   display: flex;
@@ -20,20 +22,37 @@
   justify-content: flex-end;
 }
 
-.banner {
-  display: flex;
-  justify-content: space-between;
-  margin: 24px 24px 0;
-  padding: 12px;
-  background-color: colors.$kobo-light-blue;
-  border-radius: 5px;
-  align-items: center;
+.dropzone {
+  width: 100%;
+  height: 100%;
+  position: relative;
 }
 
-.bannerIcon {
-  padding-right: 18px;
+.dropzoneOverlay {
+  display: none;
 }
 
-.bannerButton {
-  margin-left: auto;
+.dropzoneActive .dropzoneOverlay {
+  @include mixins.centerRowFlex;
+  justify-content: center;
+  flex-wrap: wrap;
+  text-align: center;
+  background-color: color.change(colors.$kobo-white, $alpha: 0.5);
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: z-indexes.$z-dropzone;
+  color: colors.$kobo-blue;
+  border: 6px dashed currentcolor;
+
+  :global {
+    h1 {
+      width: 100%;
+      margin: 6px 0 0;
+      font-size: 18px;
+      font-weight: normal;
+    }
+  }
 }
